refactor: extract config modules
